flask-rest-develop/requirements.txt
# Flask and flask extensions
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
flask-restx==1.3.0

# Server
gunicorn==22.0.0

# Tools
python-dotenv==1.0.1
